# docs/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Go-Micro-Services API",
    "description": "Go-Micro-Services is a POS API developed using microservices architecture. This API provides authentication, user management, product handling, and order processing functionalities.",
    "version": "1.0.0",
    "contact": {
      "name": "Developer",
      "email": "azzam.sykir.work@gmail.com"
    }
  },
...
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter 'Bearer {token}' to authenticate requests."
    }
  },
...
  "paths": {
...
    "/users/email/{email}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by Email",
        "operationId": "getUserByEmail",
        "description": "Get user details based on the provided Email.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "type": "string",
            "x-example": "example.user@example.com"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully by email.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "User retrieved successfully by email."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "b740369b-8957-4f36-b7ed-b0b8e1d07caf"
                    },
                    "name": {
                      "type": "string",
                      "example": "Example User"
                    },
                    "email": {
                      "type": "string",
                      "example": "example.user@example.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "$2a$10$wHCIdkxeK91Kmup0Pjt6xeu3rtqm7NXeiVbAgO5Q1z/wCkznWlTSi"
                    },
                    "balance": {
                      "type": "integer",
                      "example": 0
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T02:39:54.741293Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T02:39:54.741293Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by ID",
        "operationId": "getUserById",
        "description": "Get user details based on the provided user ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "x-example": "b740369b-8957-4f36-b7ed-b0b8e1d07caf"
          }
        ],
        "responses": {
          "200": {
            "description": "User retrieved successfully.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "User retrieved successfully."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "b740369b-8957-4f36-b7ed-b0b8e1d07caf"
                    },
                    "name": {
                      "type": "string",
                      "example": "Example User"
                    },
                    "email": {
                      "type": "string",
                      "example": "example.user@example.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "$2a$10$wHCIdkxeK91Kmup0Pjt6xeu3rtqm7NXeiVbAgO5Q1z/wCkznWlTSi"
                    },
                    "balance": {
                      "type": "integer",
                      "example": 0
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T02:39:54.741293Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T02:39:54.741293Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update user details",
        "operationId": "updateUser",
        "description": "Update user details based on the provided user ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "x-example": "b740369b-8957-4f36-b7ed-b0b8e1d07caf"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Example User"
                },
                "email": {
                  "type": "string",
                  "example": "example.user@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "$2a$10$wHCIdkxeK91Kmup0Pjt6xeu3rtqm7NXeiVbAgO5Q1z/wCkznWlTSi"
                },
                "balance": {
                  "type": "integer",
                  "example": 10000000
                }
              },
              "required": [
                "name",
                "email",
                "balance"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "User updated successfully."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "b740369b-8957-4f36-b7ed-b0b8e1d07caf"
                    },
                    "name": {
                      "type": "string",
                      "example": "Example User"
                    },
                    "email": {
                      "type": "string",
                      "example": "example.user@example.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "$2a$10$wHCIdkxeK91Kmup0Pjt6xeu3rtqm7NXeiVbAgO5Q1z/wCkznWlTSi"
                    },
                    "balance": {
                      "type": "integer",
                      "example": 10000000
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T02:39:54.741293Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T03:54:01.051294888Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user",
        "operationId": "deleteUser",
        "description": "Delete user based on the provided user ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string",
            "x-example": "b740369b-8957-4f36-b7ed-b0b8e1d07caf"
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "User with ID b740369b-8957-4f36-b7ed-b0b8e1d07caf has been successfully deleted."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "b740369b-8957-4f36-b7ed-b0b8e1d07caf"
                    },
                    "name": {
                      "type": "string",
                      "example": "Example User"
                    },
                    "email": {
                      "type": "string",
                      "example": "example.user@example.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "$2a$10$wHCIdkxeK91Kmup0Pjt6xeu3rtqm7NXeiVbAgO5Q1z/wCkznWlTSi"
                    },
                    "balance": {
                      "type": "integer",
                      "example": 10000000
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T02:39:54.741293Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T03:54:01.051295Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/categories": {
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Create a new category",
        "operationId": "createCategory",
        "description": "Registers a new category.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Example Category"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Category created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 201
                },
                "message": {
                  "type": "string",
                  "example": "Category Created Successfully"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "2111a162-a912-454d-aaf0-0feeafe60d9f"
                    },
                    "name": {
                      "type": "string",
                      "example": "example category"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T06:11:47.988013891Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T06:11:47.988013891Z"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get List Category",
        "operationId": "ListCategories",
        "description": "Retrieve a list of categories.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved category list.",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Successfully retrieved category list."
                },
                "data": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "2111a162-a912-454d-aaf0-0feeafe60d9f"
                      },
                      "name": {
                        "type": "string",
                        "example": "example category"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-03T06:11:47.988014Z"
                      },
                      "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-03T06:11:47.988014Z"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/categories/{id}": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get category by ID",
        "operationId": "getCategoryById",
        "description": "Retrieves category details based on the provided ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "x-example": "2111a162-a912-454d-aaf0-0feeafe60d9f"
          }
        ],
        "responses": {
          "200": {
            "description": "Category retrieved successfully",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Category retrieved successfully"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "2111a162-a912-454d-aaf0-0feeafe60d9f"
                    },
                    "name": {
                      "type": "string",
                      "example": "example category"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T06:11:47.988014Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T06:11:47.988014Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "summary": "Update an existing category",
        "operationId": "updateCategory",
        "description": "Updates an existing category by ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "x-example": "7c9028c2-fd8a-40e8-b42f-c6112cee382c"
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "example updated category"
                }
              },
              "required": [
                "name"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Category updated successfully."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "7c9028c2-fd8a-40e8-b42f-c6112cee382c"
                    },
                    "name": {
                      "type": "string",
                      "example": "example updated category"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T06:39:10.95694Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T06:39:12.646837852Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete a category",
        "operationId": "deleteCategory",
        "description": "Deletes a category by ID.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "x-example": "8aa955d4-3c6a-4d80-943c-6800cdbe4bdc"
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 200
                },
                "message": {
                  "type": "string",
                  "example": "Category deleted successfully."
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "8aa955d4-3c6a-4d80-943c-6800cdbe4bdc"
                    },
                    "name": {
                      "type": "string",
                      "example": "example category"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T06:34:46.547866Z"
                    },
                    "updated_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-03-03T06:34:46.547866Z"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        }
      }
    },
...
  },
...
}